requirements.txt
aiohttp==1.1.5
aiohttp-cors==0.5.0
aiohttp-swagger==1.0.2
aioredis==0.3.0
alabaster==0.7.9
apns==2.0.1
async-timeout==1.1.0
Babel==2.3.4
backpack==0.1
blinker==1.3
chardet==2.3.0
cleo==0.5.0
cloud-init==0.7.9
command-not-found==0.3
configobj==5.0.6
cryptography==1.5
cssselect==1.0.0
cssutils==1.0.1
defusedxml==0.4.1
distro-info===0.14build1
docutils==0.12
emails==0.5.10
fake-factory==0.7.2
future==0.16.0
gunicorn==19.6.0
hiredis==0.2.0
idna==2.1
imagesize==0.7.1
inflection==0.3.1
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.10
language-selector==0.1
lazy-object-proxy==1.2.2
lxml==3.6.4
MarkupSafe==0.23
mock==2.0.0
multidict==2.1.2
oauthlib==1.1.2
orator==0.9.3
pbr==1.10.0
pendulum==0.6.5
premailer==3.0.1
prettytable==0.7.2
psutil==5.0.0
psycopg2==2.6.2
py==1.4.31
pyaml==16.11.4
pyasn1==0.1.9
Pygments==2.1.3
pygobject==3.22.0
PyJWT==1.4.1
pylev==1.3.0
pyserial==3.1.1
pytest==3.0.4
python-apt==1.1.0b5
python-dateutil==2.6.0
python-debian==0.1.29
python-magic==0.4.12
python-twitter==3.2
python3-openid==3.0.10
pytz==2016.7
PyYAML==3.11
requests==2.10.0
requests-oauthlib==0.7.0
simplejson==3.10.0
six==1.10.0
snowballstemmer==1.2.1
social-auth-core==0.0.1
Sphinx==1.5
ssh-import-id==5.6
systemd-python==232
tzlocal==1.3
ufw==0.35
unattended-upgrades==0.1
urllib3==1.15.1
uvloop==0.7.2
wrapt==1.10.8
yarl==0.7.1
